Fix AddNpc NameColor type: int ARGB, not float
The client reads this field as an int32 ARGB color (m_nNameColor).
Both types are 4 bytes on the wire and the only current assignment is
default(0), so nothing breaks today - but an int color assigned through
the float conversion mangles the bits and the nameplate renders wrong.
Verified live: a static ARGB value on the int field colors the NPC's
overhead nameplate as specified.

+++ src/Sanctuary.Packet/BasePlayerUpdatePacket/PlayerUpdatePacketAddNpc.cs
@@ -103,7 +103,9 @@ public class PlayerUpdatePacketAddNpc : BasePlayerUpdatePacket, ISerializablePac
     [JsonConverter(typeof(Vector4JsonConverter))]
     public Vector4 Tilt;
 
-    public float NameColor;
+    // ARGB int32 (client m_nNameColor) — a static nameplate color; 0 lets the client's disposition
+    // resolver pick the color.
+    public int NameColor;
 
     public int AreaDefinitionId;
